+++ backend/src/test/java/peer/backend/team/InterestedProjectRepositoryTest.java
@@ -1,9 +1,8 @@
-package peer.backend.teamuser;
+package peer.backend.team;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.time.LocalDateTime;
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,29 +10,31 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import peer.backend.entity.team.Team;
-import peer.backend.entity.team.TeamUser;
 import peer.backend.entity.team.enums.TeamMemberStatus;
 import peer.backend.entity.team.enums.TeamOperationFormat;
 import peer.backend.entity.team.enums.TeamStatus;
 import peer.backend.entity.team.enums.TeamType;
+import peer.backend.entity.user.InterestedProject;
 import peer.backend.entity.user.User;
 import peer.backend.repository.team.TeamRepository;
-import peer.backend.repository.team.TeamUserRepository;
+import peer.backend.repository.user.InterestedProjectRepository;
 import peer.backend.repository.user.UserRepository;
 
-@DisplayName("Team User Repository 테스트")
+@DisplayName("Interested Project Repository 테스트")
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class TeamUserRepositoryTest {
+public class InterestedProjectRepositoryTest {
 
     @Autowired
-    TeamUserRepository teamUserRepository;
+    InterestedProjectRepository interestedProjectRepository;
 
     @Autowired
     UserRepository userRepository;
 
     @Autowired
     TeamRepository teamRepository;
+
+    InterestedProject interestedProject;
 
     @BeforeEach
     void beforeEach() {
@@ -69,75 +70,57 @@ public class TeamUserRepositoryTest {
     }
 
     @Test
-    @DisplayName("TeamUser insert test")
+    @DisplayName("Interested Project insert test")
     void insertTest() {
         User user = userRepository.findAll().get(0);
         Team team = teamRepository.findAll().get(0);
 
-        TeamUser teamUser = TeamUser.builder()
+        InterestedProject interestedProject = InterestedProject.builder()
             .user(user)
             .userId(user.getId())
             .team(team)
             .teamId(team.getId())
             .build();
 
-        teamUserRepository.save(teamUser);
-        assertEquals(teamUserRepository.count(), 1);
+        interestedProjectRepository.save(interestedProject);
+        assertEquals(interestedProjectRepository.count(), 1);
     }
 
     @Test
-    @DisplayName("TeamUser findByUserIdAndTeamId test")
-    void findByUserIdAndTeamIdTest() {
+    @DisplayName("Interested Project select test")
+    void selectTest() {
         User user = userRepository.findAll().get(0);
         Team team = teamRepository.findAll().get(0);
 
-        TeamUser teamUser = TeamUser.builder()
+        InterestedProject interestedProject = InterestedProject.builder()
             .user(user)
             .userId(user.getId())
             .team(team)
             .teamId(team.getId())
             .build();
 
-        teamUserRepository.save(teamUser);
-        TeamUser find = teamUserRepository.findByUserIdAndTeamId(user.getId(),
+        interestedProjectRepository.save(interestedProject);
+        InterestedProject find = interestedProjectRepository.findByUserIdAndTeamId(user.getId(),
             team.getId());
-        assertEquals(find.getTeamId(), teamUser.getTeamId());
+        assertEquals(find.getTeamId(), interestedProject.getTeamId());
     }
 
     @Test
-    @DisplayName("TeamUser findByUserId test")
-    void findByUserIdTest() {
-        User user = userRepository.findAll().get(0);
-        Team team = teamRepository.findAll().get(0);
-
-        TeamUser teamUser = TeamUser.builder()
-            .user(user)
-            .userId(user.getId())
-            .team(team)
-            .teamId(team.getId())
-            .build();
-
-        teamUserRepository.save(teamUser);
-        List<TeamUser> teamUserList = teamUserRepository.findByUserId(user.getId());
-        assertEquals(teamUserList.size(), 1);
-    }
-
-    @Test
-    @DisplayName("TeamUser delete test")
+    @DisplayName("Interested Project delete test")
     void deleteTest() {
         User user = userRepository.findAll().get(0);
         Team team = teamRepository.findAll().get(0);
 
-        TeamUser teamUser = TeamUser.builder()
+        InterestedProject interestedProject = InterestedProject.builder()
             .user(user)
             .userId(user.getId())
             .team(team)
             .teamId(team.getId())
             .build();
 
-        teamUserRepository.save(teamUser);
-        assertEquals(teamUserRepository.count(), 1);
-        teamUserRepository.deleteByUserIdAndTeamId(user.getId(), team.getId());
-        assertEquals(teamUserRepository.count(), 0);
+        interestedProjectRepository.save(interestedProject);
+        assertEquals(interestedProjectRepository.count(), 1);
+        interestedProjectRepository.deleteByUserIdAndTeamId(user.getId(), team.getId());
+        assertEquals(interestedProjectRepository.count(), 0);
     }
 }
